move some operations to main
